Create How to Check the Installation Location of a Tool on a Linux Server
# How to Check the Installation Location of a Tool on a Linux Server

📋 Introduction

When working with a Linux server, it's crucial to know where specific tools are installed. In this comprehensive guide, we will walk you through various methods to determine the installation location of a tool on your Linux server. We will use popular tool Jenkins as an example, but you can apply these techniques to any tool of your choice.